Add support for filtering names using substrings. Add colors. (#7)
* Add support for filtering names using regexps. Add colors.

* Disable linting on similar functions. Remove unused colour.

* Bump Go version. Put the linter ignore at the top.

* Change to support based on substrings.

* Bump Goreleaser version.

+++ internal/civo/networks.go
@@ -1,40 +1,47 @@
+//nolint:dupl // similar functions due to upstream packaging
 package civo
 
 import (
 	"fmt"
 
+	"github.com/konstructio/dropkick/internal/compare"
 	"github.com/konstructio/dropkick/internal/outputwriter"
 )
 
 // NukeNetworks deletes all networks associated with the Civo client.
 // It returns an error if the deletion process encounters any issues.
 func (c *Civo) NukeNetworks() error {
-	c.logger.Printf("listing networks")
+	c.logger.Infof("listing networks")
 
 	networks, err := c.client.ListNetworks()
 	if err != nil {
 		return fmt.Errorf("unable to list networks: %w", err)
 	}
 
-	c.logger.Printf("found %d networks", len(networks))
+	c.logger.Infof("found %d networks", len(networks))
 
 	for _, network := range networks {
-		c.logger.Printf("found network: name: %q - ID: %q", network.Name, network.ID)
+		c.logger.Infof("found network: name: %q - ID: %q", network.Name, network.ID)
+
+		if c.nameFilter != "" && !compare.ContainsIgnoreCase(network.Name, c.nameFilter) {
+			c.logger.Warnf("skipping network %q: name does not match filter", network.Name)
+			continue
+		}
 
 		if c.nuke {
-			c.logger.Printf("deleting network %q", network.ID)
+			c.logger.Infof("deleting network %q", network.Name)
 			res, err := c.client.DeleteNetwork(network.ID)
 			if err != nil {
-				return fmt.Errorf("unable to delete network %q: %w", network.ID, err)
+				return fmt.Errorf("unable to delete network %q: %w", network.Name, err)
 			}
 
 			if res.ErrorCode != "200" {
-				return fmt.Errorf("Civo returned an error code %s when deleting network %q: %s", res.ErrorCode, network.ID, res.ErrorDetails)
+				return fmt.Errorf("Civo returned an error code %q when deleting network %q: %s", res.ErrorCode, network.Name, res.ErrorDetails)
 			}
 
-			outputwriter.WriteStdoutf("deleted network %q", network.ID)
+			outputwriter.WriteStdoutf("deleted network %q", network.Name)
 		} else {
-			c.logger.Printf("refusing to delete network %q: nuke is not enabled", network.ID)
+			c.logger.Warnf("refusing to delete network %q: nuke is not enabled", network.Name)
 		}
 	}
